Use a redirectUrl prop from EntityPageDocs to inject the redirection of search results
