+++ plugins/techdocs/src/reader/components/TechDocsReaderPageContent/TechDocsReaderPageContent.tsx
@@ -66,10 +66,11 @@ export type TechDocsReaderPageContentProps = {
   withSearch?: boolean;
   /**
    * If {@link TechDocsReaderPageContentProps.withSearch | withSearch} is true,
-   * this will turn search results into links within the "Docs" tab of the
-   * entity page, instead of the global docs page.
+   * this will redirect the search result urls, and turn search results into
+   * links within the "Docs" tab of the entity page, instead of the global docs
+   * page.
    */
-  redirectToEntityPage?: boolean;
+  redirectUrl?: (url: string) => string;
   /**
    * Callback called when the content is rendered.
    */
@@ -82,7 +83,7 @@ export type TechDocsReaderPageContentProps = {
  */
 export const TechDocsReaderPageContent = withTechDocsReaderProvider(
   (props: TechDocsReaderPageContentProps) => {
-    const { withSearch = true, redirectToEntityPage = false, onReady } = props;
+    const { withSearch = true, redirectUrl, onReady } = props;
     const classes = useStyles();
 
     const {
@@ -148,7 +149,7 @@ export const TechDocsReaderPageContent = withTechDocsReaderProvider(
               <TechDocsSearch
                 entityId={entityRef}
                 entityTitle={entityMetadata?.metadata?.title}
-                redirectToEntityPage={redirectToEntityPage}
+                redirectUrl={redirectUrl}
               />
             </Grid>
           )}
