+++ plugins/techdocs/src/search/components/TechDocsSearch.tsx
@@ -24,7 +24,6 @@ import {
 import React, { useEffect, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { TechDocsSearchResultListItem } from './TechDocsSearchResultListItem';
-import { useTechDocsLocation } from './useTechDocsLocation';
 
 /**
  * Props for {@link TechDocsSearch}
@@ -35,7 +34,7 @@ export type TechDocsSearchProps = {
   entityId: CompoundEntityRef;
   entityTitle?: string;
   debounceTime?: number;
-  redirectToEntityPage?: boolean;
+  redirectUrl?: (url: string) => string;
 };
 
 type TechDocsDoc = {
@@ -60,12 +59,7 @@ const isTechDocsSearchResult = (
 };
 
 const TechDocsSearchBar = (props: TechDocsSearchProps) => {
-  const {
-    entityId,
-    entityTitle,
-    debounceTime = 150,
-    redirectToEntityPage = false,
-  } = props;
+  const { entityId, entityTitle, debounceTime = 150, redirectUrl } = props;
   const [open, setOpen] = useState(false);
   const navigate = useNavigate();
   const {
@@ -103,15 +97,13 @@ const TechDocsSearchBar = (props: TechDocsSearchProps) => {
     });
   }, [kind, namespace, name, setFilters]);
 
-  const techDocsLocation = useTechDocsLocation(entityId, redirectToEntityPage);
-
   const handleSelection = (
     _: any,
     selection: TechDocsSearchResult | string | null,
   ) => {
     if (isTechDocsSearchResult(selection)) {
       const { location } = selection.document;
-      navigate(techDocsLocation(location));
+      navigate(redirectUrl ? redirectUrl(location) : location);
     }
   };
